+++ apps/play.strawberry.rocks/src/app.tsx
@@ -1,141 +1,7 @@
-import { CodeEditor } from "./components/editor/editor";
-import { ResizeHandler } from "./components/resize-handler";
-import { StatusBadge } from "./components/status-badge";
+import { Playground } from "./components/playground";
 import { usePyodide } from "./components/strawberry/pyodide";
-import { Tabs, Tab } from "./components/tabs";
 import { VersionSelector } from "./components/version-selector";
-import { useState, useEffect, Suspense, useCallback } from "react";
-import { Panel, PanelGroup } from "react-resizable-panels";
-
-const STARTER_CODE = `
-import strawberry
-
-@strawberry.type
-class Query:
-    @strawberry.field
-    def hello(self, info: strawberry.Info) -> str:
-        return "world"
-
-schema = strawberry.Schema(Query)
-`.trim();
-
-type Result = {
-  data: any;
-  status_code: number;
-  headers: { [key: string]: string };
-};
-
-const Playground = () => {
-  const { executeQuery } = usePyodide();
-  const [editorState, setEditorState] = useState({
-    code: STARTER_CODE,
-    query: "{ hello }",
-    variables: "{}",
-  });
-
-  const [result, setResult] = useState<Result | null>(null);
-
-  const runQuery = useCallback(async () => {
-    const code = editorState.code;
-    const query = editorState.query;
-    const variables = editorState.variables;
-
-    const { result, error } = await executeQuery({
-      schemaCode: code,
-      query,
-      variables,
-    });
-
-    // TODO: handle error?
-
-    setResult(result);
-  }, [editorState]);
-
-  useEffect(() => {
-    console.log("running");
-    runQuery();
-  }, [editorState.code, editorState.query, editorState.variables]);
-
-  return (
-    <PanelGroup direction="horizontal" className="h-full">
-      <Panel minSize={20}>
-        <Tabs className="h-full">
-          <Tab title="Code">
-            <CodeEditor
-              source={editorState.code}
-              onChange={(code) => {
-                setEditorState({ ...editorState, code });
-              }}
-              language="python"
-            />
-          </Tab>
-        </Tabs>
-      </Panel>
-      <ResizeHandler />
-
-      <Panel minSize={20}>
-        <PanelGroup direction="vertical" className="h-full">
-          <Panel minSize={20}>
-            <Tabs className="h-full">
-              <Tab title="Query">
-                <CodeEditor
-                  source={editorState.query}
-                  onChange={(query) => {
-                    setEditorState({ ...editorState, query });
-                  }}
-                  language="graphql"
-                />
-              </Tab>
-            </Tabs>
-          </Panel>
-          <ResizeHandler direction="vertical" />
-          <Panel minSize={20}>
-            <Tabs className="h-full">
-              <Tab title="Variables">
-                <CodeEditor
-                  source={editorState.variables}
-                  onChange={(variables) => {
-                    setEditorState({ ...editorState, variables });
-                  }}
-                  language="json"
-                />
-              </Tab>
-            </Tabs>
-          </Panel>
-        </PanelGroup>
-      </Panel>
-      <ResizeHandler />
-      <Panel minSize={20}>
-        <div className="h-full flex flex-col">
-          <Tabs className="flex-1 w-full">
-            <Tab title="Result">
-              <CodeEditor
-                source={
-                  result?.data ? JSON.stringify(result.data, null, 2) : ""
-                }
-                language="json"
-                readOnly
-              />
-            </Tab>
-            <Tab title="Headers">
-              <CodeEditor
-                source={
-                  result?.headers ? JSON.stringify(result.headers, null, 2) : ""
-                }
-                language="json"
-                readOnly
-              />
-            </Tab>
-          </Tabs>
-
-          <div className="p-2 border-t">
-            <StatusBadge status={result?.status_code ?? 200} />
-          </div>
-        </div>
-      </Panel>
-    </PanelGroup>
-  );
-};
+import { useState, Suspense } from "react";
 
 function App() {
   const { initialLoading } = usePyodide();
